refactor(message-utils): centralize format_prefix_messages_content function
Move the format_prefix_messages_content function from main.py and streamlit_chat.py to message_utils.py to centralize its definition and avoid code duplication.
